+++ src/importlinter/configuration.py
@@ -1,6 +1,6 @@
 from .adapters.building import GraphBuilder
 from .adapters.filesystem import FileSystem
-from .adapters.printing import ClickPrinter
+from .adapters.printing import RichPrinter
 from .adapters.timing import SystemClockTimer
 from .adapters.user_options import IniFileUserOptionReader, TomlFileUserOptionReader
 from .application.app_config import settings
@@ -13,7 +13,7 @@ def configure():
             "toml": TomlFileUserOptionReader(),
         },
         GRAPH_BUILDER=GraphBuilder(),
-        PRINTER=ClickPrinter(),
+        PRINTER=RichPrinter(),
         FILE_SYSTEM=FileSystem(),
         TIMER=SystemClockTimer(),
         DEFAULT_CACHE_DIR=".import_linter_cache",